+++ src/test/java/nz/co/searchwellington/controllers/models/GeotaggedModelBuilderTest.java
@@ -27,7 +27,7 @@ import uk.co.eelpieconsulting.common.geo.model.Place;
 public class GeotaggedModelBuilderTest {
 	
 	private static final long TOTAL_GEOTAGGED_COUNT = 512;
-	private static final int LOCATION_RESULTS_COUNT = 33;
+	private static final long LOCATION_RESULTS_COUNT = 33;
 	
 	@Mock ContentRetrievalService contentRetrievalService;
 	@Mock UrlBuilder urlBuilder;
